+++ SamyuktaManabSoftwares/GrocerySupplyManagementApp/Forms/SupplierForm.cs
@@ -319,16 +319,20 @@ namespace GrocerySupplyManagementApp.Forms
 
         private void BtnDelete_Click(object sender, EventArgs e)
         {
-            var supplierName = TxtSupplierName.Text;
-            _supplierService.DeleteSupplier(supplierName);
-
-            DialogResult result = MessageBox.Show(supplierName + " has been deleted successfully.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            if (result == DialogResult.OK)
+            DialogResult deleteResult = MessageBox.Show(Constants.MESSAGE_BOX_DELETE_MESSAGE, "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (deleteResult == DialogResult.Yes)
             {
-                ClearAllFields();
-                DataGridSupplierList.DataSource = null;
-                EnableFields();
-                EnableFields(Action.DeleteSupplier);
+                var supplierName = TxtSupplierName.Text;
+                _supplierService.DeleteSupplier(supplierName);
+
+                DialogResult result = MessageBox.Show(supplierName + " has been deleted successfully.", "Information", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                if (result == DialogResult.OK)
+                {
+                    ClearAllFields();
+                    DataGridSupplierList.DataSource = null;
+                    EnableFields();
+                    EnableFields(Action.DeleteSupplier);
+                }
             }
         }
